requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.8
discord.py==2.3.2
flake8==7.0.0
frozenlist==1.3.3
iniconfig==2.0.0
langchain==0.0.220
langchainplus-sdk==0.0.17
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.0
openapi-schema-pydantic==1.2.4
packaging==23.1
pluggy==1.5.0
pycodestyle==2.11.1
pydantic==1.10.10
pyflakes==3.2.0
pytest==8.2.1
python-dotenv==1.0.1
PyYAML==6.0
requests==2.31.0
SQLAlchemy==2.0.17
tenacity==8.2.2
typing-inspect==0.9.0
typing_extensions==4.7.0
urllib3==2.0.3
yarl==1.9.2